+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable, unreleased changes to this project will be documented in this file.
 
 
 - Add our implementation of UUID scalar - #5646 by @koradon
+- Prevent creating checkout/draft order with unpublished product - #5676 by @d-wysocki
 
 ## 2.10.0